requirements.txt
certifi>=2022.6.15
chardet>=5.0.0
idna>=3.3
requests>=2.28.1
todoist-python>=8.1.3
urllib3>=1.26.12
datetime>=4.3